requirements.txt
flask==2.3.0
flask-cors==4.0.0
Werkzeug==2.3.0
picamera2==0.3.9
adafruit-pca9685==1.4.1
RPi.GPIO==0.7.0
